Made some big changes.
I have removed the old blogpostapp django app. The migrations were having issues that were not resolvable in my limited knowledge of github and django.

In place of the old blogpostapp I am using ElectronixBlogPDX it is mostly the same with a few changes in the models.
Also I have removed all files that are not related to the new blog. I am downsizing the project to just a blog and and the posts I make. The future may see an amazing blog with more going on. I would like to have users who can post comments and projects and login into the community chat page. This is no longer possible with the time frame the project has.