9-15-15
This is on the way to good progress.
I've got the
HTML homepage getting called
CSS is working on the page
base html and static-styles.css loading properly

the forms have the right fields
the forms are saving data
:)
:)
:)

I will next work on the viewing of the blogpost getting pulled to the page and the forms being viewed in the page as well.
_______________________
9-16-15

I have a better looking  page with a lot of bootstrap features.
Still need to work on getting the  forms to display in the page and have functionality.

___________________
9-17-15
So I have the page loading the bootstrap but the css is not found now. I am able to link to the page from the Admin site in django.
I am still needing to add the forms to the page and the pull the post to the page, the latest post being found on the homepage and the other post being archived in the blog posts past page(to be made).