Lots of good changes,
updated models class name so that it is unique, not the same as the class in forms.

added a number of html pages

updated admin, urls, views , and forms to reflect the changes to the class name change in models and the additional pages.

still trying to straighten out the bumps, as much of the changes are based on (not copied from) work in the tutorials I 've worked through.