I have been working hard to line up project and pycharms tools that have been making it hard to see where my own errors were poping up.
I've id'ed the different types of files in the project. I've gone through most all references of views and html rewriting them so that pycharm is able to recognise them.

Over all the page is starting to work like it should. There is a problem with the page when it jumps to a new html, the css is linked to the base and maybe should be tied into each individual page.

just got the static link in base.html to work like it should, just took weeks of trying different things.