Got some help with the project from Chris Jones at PDX Code Guild. We were able to resolve the issue that I'd been having with the development server in django. The  issue was a result of the last push to GitHub. I had been asked to rebase the project or merge the files between the repository and local. In doing the merge the manage.py file had been changed. Line 6 had changed from "electronix.settings" to "electronix_pdx.settings", a quick change and things are working better now.
Another step forward was made with the help of another coder here, who was able to show what has been causing the css/bootstrap to hangup when leaving the main page.

First I needed to add {% load staticfiles%}  to each of the pages that I have in the templates directory.

Second the settings file was missing a small detail, the STATICFILES_DIR needed to have the full directory listed to point to the static folder (electronix_pdx/static) this is needed only because I have the static folder under the app directory, not directly under the project folder.

This two things have real helped.
Know to work out the issues with the urls and views. :(