Great improvements today. Worked out an number of issues that have been stalling this project. With Chris's help was abel to resolve the broken nav links. I had tried to use a straight link to the HTML page needed. Chris should how to use the views and URLS files to reach the right page. Then we setup the page to connect to the right links. The homepage was not displaying the latest post, showing all of them. He showed me how to write the view that was needed to display the latest.
I was able to figure out the reason for only the edit link/glyphicon showing on the single blog pages. I had forgotten to create the link for the remove button/glyphicon.

Another issue resolved tonight was the problem with static files not getting resolved. The settings file had been looking at the static section as a list and it need to be writen as a tuple(or something of that nature). Somehow by adding parens. to the front and end of the os. ...it became a tuple.